add merge welcome screen with the new design
